Add ability to reloeload,cancel, upload

+++ app/src/main/java/com/worldventures/dreamtrips/modules/bucketlist/api/UploadBucketPhotoCommand.java
@@ -6,6 +6,7 @@ import com.techery.spares.module.Injector;
 import com.worldventures.dreamtrips.core.api.request.DreamTripsRequest;
 import com.worldventures.dreamtrips.core.repository.SnappyRepository;
 import com.worldventures.dreamtrips.core.utils.events.UploadProgressUpdateEvent;
+import com.worldventures.dreamtrips.modules.bucketlist.event.BucketPhotoUploadCancelEvent;
 import com.worldventures.dreamtrips.modules.bucketlist.event.BucketPhotoUploadFailedEvent;
 import com.worldventures.dreamtrips.modules.bucketlist.event.BucketPhotoUploadFinished;
 import com.worldventures.dreamtrips.modules.bucketlist.event.BucketPhotoUploadStarted;
@@ -48,17 +49,20 @@ public class UploadBucketPhotoCommand extends DreamTripsRequest<BucketPhoto> {
             int taskId = photoUploadTask.getTaskId();
 
             String urlFromUploadResult = s3uploader.uploadImageToS3(fileUri, String.valueOf(taskId));
+
             BucketPhoto uploadObject = getUploadObject(taskId, urlFromUploadResult);
 
-            eventBus.post(new UploadProgressUpdateEvent(String.valueOf(taskId), 100));
+            if (!isCancelled()) {
+                BucketPhoto photo = getService().uploadBucketPhoto(photoUploadTask.getBucketId(), uploadObject);
+                photo.setTaskId(taskId);
 
-            BucketPhoto photo = getService().uploadBucketPhoto(photoUploadTask.getBucketId(), uploadObject);
-            photo.setTaskId(taskId);
+                eventBus.post(new UploadProgressUpdateEvent(String.valueOf(taskId), 100));
 
+              //  eventBus.post(new BucketPhotoUploadFinished(photo));
+                return photo;
+            }
+            eventBus.post(new BucketPhotoUploadCancelEvent(photoUploadTask));
             db.removeBucketPhotoTask(photoUploadTask);
-            eventBus.post(new BucketPhotoUploadFinished(photo));
-
-            return photo;
         } catch (Exception e) {
             Log.e(this.getClass().getSimpleName(), e);
             eventBus.post(new BucketPhotoUploadFailedEvent(photoUploadTask.getTaskId()));
